chore: migrate contact service to version 2 and update related references
