+++ bc_obps/registration/api/_operations/_operation_id/operation_representatives.py
@@ -8,7 +8,7 @@ from common.api.utils import get_current_user_guid
 from registration.constants import OPERATION_TAGS
 from registration.models.contact import Contact
 from registration.schema import OperationRepresentativeListOut, Message
-from service.contact_service import ContactService
+from service.contact_service_v2 import ContactServiceV2
 from service.error_service.custom_codes_4xx import custom_codes_4xx
 
 
@@ -25,4 +25,4 @@ def list_operation_representatives(
     request: HttpRequest,
     operation_id: UUID,
 ) -> QuerySet[Contact]:
-    return ContactService.list_operation_representatives(operation_id, get_current_user_guid(request))
+    return ContactServiceV2.list_operation_representatives(operation_id, get_current_user_guid(request))
